refactor: move interpreting expression logic to interpreter folder
