+++ core/engine/src/lib.rs
@@ -6,6 +6,7 @@ pub mod module;
 pub mod natives;
 pub mod value;
 pub mod vm;
+mod interpreter;
 
 pub use roan_ast::*;
 pub use roan_error::{diagnostic::*, error::PulseError::*, span::*};
